pull from original

+++ app/Http/Controllers/EventSurveyController.php
@@ -8,6 +8,7 @@ use App\Models\Attendee;
 use App\Models\Question;
 use App\Models\QuestionAnswer;
 use App\Models\QuestionType;
+use JavaScript;
 use Illuminate\Http\Request;
 use Excel;
 
@@ -17,10 +18,9 @@ use Excel;
 
 class EventSurveyController extends MyBaseController
 {
-    
     /**
      * Show the event survey page
-     * 
+     *
      * @param Request $request
      * @param $event_id
      * @return mixed
@@ -29,12 +29,16 @@ class EventSurveyController extends MyBaseController
     {
         $event = Event::scope()->findOrFail($event_id);
 
+        JavaScript::put([
+            'postUpdateQuestionsOrderRoute' => route('postUpdateQuestionsOrder', ['event_id' => $event_id]),
+        ]);
+
         $data = [
-            'event' => $event,
-            'questions' => $event->questions,
+            'event'      => $event,
+            'questions'  => $event->questions->sortBy('sort_order'),
             'sort_order' => 'asc',
-            'sort_by' => 'title',
-            'q' => '',
+            'sort_by'    => 'title',
+            'q'          => '',
         ];
 
         return view('ManageEvent.Surveys', $data);
@@ -50,7 +54,7 @@ class EventSurveyController extends MyBaseController
         $event = Event::scope()->findOrFail($event_id);
 
         return view('ManageEvent.Modals.CreateQuestion', [
-            'event' => $event,
+            'event'          => $event,
             'question_types' => QuestionType::all(),
         ]);
     }
@@ -60,18 +64,17 @@ class EventSurveyController extends MyBaseController
      *
      * @access public
      * @param  StoreEventQuestionRequest $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
     public function postCreateEventQuestion(StoreEventQuestionRequest $request, $event_id)
     {
-
         // Get the event or display a 'not found' warning.
         $event = Event::findOrFail($event_id);
 
         // Create question.
         $question = Question::createNew(false, false, true);
         $question->title = $request->get('title');
-        $question->is_required = ($request->get('is_required') == 'yes') ;
+        $question->is_required = ($request->get('is_required') == 'yes');
         $question->question_type_id = $request->get('question_type_id');
         $question->save();
 
@@ -99,8 +102,8 @@ class EventSurveyController extends MyBaseController
         session()->flash('message', 'Successfully Created Question');
 
         return response()->json([
-            'status' => 'success',
-            'message' => 'Refreshing..',
+            'status'      => 'success',
+            'message'     => 'Refreshing..',
             'redirectUrl' => '',
         ]);
     }
@@ -108,7 +111,7 @@ class EventSurveyController extends MyBaseController
 
     /**
      * Show the Edit Question Modal
-     * 
+     *
      * @param Request $request
      * @param $event_id
      * @param $question_id
@@ -120,8 +123,8 @@ class EventSurveyController extends MyBaseController
         $event = Event::scope()->findOrFail($event_id);
 
         $data = [
-            'question' => $question,
-            'event' => $event,
+            'question'       => $question,
+            'event'          => $event,
             'question_types' => QuestionType::all(),
         ];
 
@@ -130,12 +133,12 @@ class EventSurveyController extends MyBaseController
 
 
     /**
-     * Edits a question
-     * 
+     * Edit a question
+     *
      * @param Request $request
      * @param $event_id
      * @param $question_id
-     * @return mixed
+     * @return \Illuminate\Http\JsonResponse
      */
     public function postEditEventQuestion(Request $request, $event_id, $question_id)
     {
@@ -149,32 +152,37 @@ class EventSurveyController extends MyBaseController
         $question->question_type_id = $request->get('question_type_id');
         $question->save();
 
-        // Get options.
-        $options = $request->get('option');
+        $question_type = QuestionType::find($question->question_type_id);
 
-        $question->options()->delete();
+        if ($question_type->has_options) {
 
-        // Add options.
-        if ($options && is_array($options)) {
-            foreach ($options as $option_name) {
-                if (trim($option_name) !== '') {
-                    $question->options()->create([
-                        'name' => $option_name,
-                    ]);
+            // Get options.
+            $options = $request->get('option');
+
+            $question->options()->delete();
+
+            // Add options.
+            if ($options && is_array($options)) {
+                foreach ($options as $option_name) {
+                    if (trim($option_name) !== '') {
+                        $question->options()->create([
+                            'name' => $option_name,
+                        ]);
+                    }
                 }
             }
         }
 
         // Get tickets.
-        $ticket_ids = $request->get('tickets');
+        $ticket_ids = (array)$request->get('tickets');
 
         $question->tickets()->sync($ticket_ids);
 
         session()->flash('message', 'Successfully Edited Question');
 
         return response()->json([
-            'status' => 'success',
-            'message' => 'Refreshing..',
+            'status'      => 'success',
+            'message'     => 'Refreshing..',
             'redirectUrl' => '',
         ]);
 
@@ -182,10 +190,10 @@ class EventSurveyController extends MyBaseController
 
     /**
      * Delete a question
-     * 
+     *
      * @param Request $request
      * @param $event_id
-     * @return mixed
+     * @return \Illuminate\Http\JsonResponse
      */
     public function postDeleteEventQuestion(Request $request, $event_id)
     {
@@ -193,14 +201,16 @@ class EventSurveyController extends MyBaseController
 
         $question = Question::scope()->find($question_id);
 
+        $question->answers()->delete();
+
         if ($question->delete()) {
 
             session()->flash('message', 'Question Successfully Deleted');
 
             return response()->json([
-                'status'  => 'success',
-                'message' => 'Refreshing..',
-                'redirectUrl' => route('showEventCustomize', ['event_id' => $event_id]) . '#questions',
+                'status'      => 'success',
+                'message'     => 'Refreshing..',
+                'redirectUrl' => '',
             ]);
         }
 
@@ -238,9 +248,16 @@ class EventSurveyController extends MyBaseController
     }
 
 
-    public function showExportAnswers(Request $request, $event_id, $export_as = 'xlsx') {
-
-        Excel::create('answers-as-of-'.date('d-m-Y-g.i.a'), function ($excel) use ($event_id) {
+    /**
+     * Export answers to xls, csv etc.
+     *
+     * @param Request $request
+     * @param $event_id
+     * @param string $export_as
+     */
+    public function showExportAnswers(Request $request, $event_id, $export_as = 'xlsx')
+    {
+        Excel::create('answers-as-of-' . date('d-m-Y-g.i.a'), function ($excel) use ($event_id) {
 
             $excel->setTitle('Survey Answers');
 
@@ -248,15 +265,11 @@ class EventSurveyController extends MyBaseController
             $excel->setCreator(config('attendize.app_name'))
                 ->setCompany(config('attendize.app_name'));
 
-            $excel->sheet('survey_answers_sheet_!', function ($sheet) use ($event_id) {
+            $excel->sheet('survey_answers_sheet_', function ($sheet) use ($event_id) {
 
                 $event = Event::scope()->findOrFail($event_id);
 
-                /*
-                 * @todo Build array of question and all its answers
-                 */
-
-                $sheet->fromArray([]);
+                $sheet->fromArray($event->survey_answers, null, 'A1', false, false);
 
                 // Set gray background on first row
                 $sheet->row(1, function ($row) {
@@ -264,9 +277,59 @@ class EventSurveyController extends MyBaseController
                 });
             });
         })->export($export_as);
+    }
 
+    /**
+     * Toggle the enabled status of question
+     *
+     * @param Request $request
+     * @param $event_id
+     * @param $question_id
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function postEnableQuestion(Request $request, $event_id, $question_id)
+    {
+        $question = Question::scope()->find($question_id);
 
+        $question->is_enabled = ($question->is_enabled == 1) ? 0 : 1;
+
+        if ($question->save()) {
+            return response()->json([
+                'status'  => 'success',
+                'message' => 'Question Successfully Updated',
+                'id'      => $question->id,
+            ]);
+        }
+
+        return response()->json([
+            'status'  => 'error',
+            'id'      => $question->id,
+            'message' => 'Whoops! Looks like something went wrong. Please try again.',
+        ]);
     }
 
 
+    /**
+     * Updates the sort order of event questions
+     *
+     * @param Request $request
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function postUpdateQuestionsOrder(Request $request)
+    {
+        $question_ids = $request->get('question_ids');
+        $sort = 1;
+
+        foreach ($question_ids as $question_id) {
+            $question = Question::scope()->find($question_id);
+            $question->sort_order = $sort;
+            $question->save();
+            $sort++;
+        }
+
+        return response()->json([
+            'status'  => 'success',
+            'message' => 'Question Order Successfully Updated',
+        ]);
+    }
 }
